delete vote and execute

+++ config.go
@@ -11,7 +11,7 @@ type Config struct {
 	TelegramAPIKey string `json:"telegram_api_key"`
 	Dev            bool   `json:"dev"`
 	LogAccount     int64  `json:"log_account"`
-	Votes          int    `json:"votes"`
+	Votes          int64  `json:"votes"`
 	Dels           int64  `json:"dels"`
 }
